ADD AMC SDT build flow * Add SDT flow for plm interface * Replaced xsct with empyro

+++ fw/AMC/src/device_drivers/ospi/ospi.h
@@ -22,7 +22,7 @@
  */
 typedef struct OSPI_CFG_TYPE
 {
-    uint8_t ucDeviceId;
+    uint32_t ucDeviceId;
     uint16_t usPageSize;
 
 } OSPI_CFG_TYPE;
