+++ fw/AMC/src/fal/ospi/fw_if_ospi.h
@@ -80,7 +80,7 @@ typedef enum FW_IF_OSPI_STATE
  */
 typedef struct FW_IF_OSPI_INIT_CFG
 {
-    uint8_t     ucOspiDeviceId;
+    uint32_t    ucOspiDeviceId;
     uint16_t    usPageSize;
 
 } FW_IF_OSPI_INIT_CFG;
